added update vehicle api:

+++ src/config/db.ts
@@ -25,7 +25,7 @@ const initDB = async () => {
         vehicle_name VARCHAR(100) NOT NULL,
         type VARCHAR(50) NOT NULL CHECK(type IN ('car', 'bike', 'van', 'SUV')),
         registration_number VARCHAR(50) UNIQUE NOT NULL,
-        daily_rent_price NUMERIC(10, 2) NOT NULL CHECK(daily_rent_price > 0),
+        daily_rent_price INTEGER NOT NULL CHECK(daily_rent_price > 0),
         availability_status VARCHAR(20) NOT NULL DEFAULT 'available' CHECK(availability_status IN ('available', 'booked'))
     );`
     )
@@ -37,7 +37,7 @@ const initDB = async () => {
         vehicle_id INTEGER NOT NULL REFERENCES vehicles(id) ON DELETE CASCADE,
         rent_start_date DATE NOT NULL,
         rent_end_date DATE NOT NULL CHECK(rent_end_date > rent_start_date),
-        total_price NUMERIC(10, 2) NOT NULL CHECK(total_price > 0),
+        total_price INTEGER NOT NULL CHECK(total_price > 0),
         status VARCHAR(20) NOT NULL CHECK(status IN ('active', 'cancelled', 'returned'))
     );`
     )
